Problems 4-6
For this work session, I attempted problems 4-6. I am pretty confident for my solutions on 4 and 6, but I still need to work on 5 some more, as I am kind of confused on the solution for finding the lcm (I tried to use the gcd function inside of python to get it, but I was unsuccessful.) I also started commenting on the code that I wrote, so it will save me time at the end of this lab.