problem 8 start
I switched to a new computer today, and was unable to get visual studio code working properly. I spent much longer than the allotted time today trying to fix that, and I was still unsuccessful. I still wanted to work a little on the project, so I tried to solve problem 8 the best I could with my resources. I was trying to split the long number into 13 equal segments in a list, but it kept creating a long list with each number as it's index, this is not what I wanted, but may be useful for the rest of the problem as I will after to multiply these numbers together.