Problem 2 Start
Today my plan was to go back and fix the past problems I couldn't get to. I started with 2 because that was the one I was the most unsure about. I changed my method for this one, opting to go for a function, which in retrospect, may end up being more complex than it needs to be. I ran out of time before I was able to finish, mainly because I am having trouble connecting my helper function which displays the fib sequence to the function that will find the even values in that sequence. Hopefully, with a bit more time I can improve this.