Problem 5
Today I solved problem 5. I spent most of the time trying to see if my old solution using the least common multiples would work, but I gave up on that. I ended up going with a helper function that goes through numbers and checks if they are divisible and then uses a counter to iterate through numbers until it finds a number that is divisible by all the numbers in the range of 1-20.  The solution took a while to figure out, and I spent over the time required to solve it.