Work Session 6: January 24th 2021 (6-8PM)
Today I finished the last assignment, focusing on problems 1, 2, and 8. I was able to figure out both 1-2, but I was stuck on 8 for a little bit. I discussed with Andy Zhang about possible solutions and he helped me understand the basic concept of recursions and said that it would be the most optimal solution for what I am trying to do. He also guided me through parts of the code, such as the project function, however, this is still my work. Looking back at the guidelines for the project, I am worried that I might have submitted things incorrectly. To clarify, I did complete 6 work sessions on time and worked for 75+ minutes on each day. For today, Sunday the 24th, I worked between 6 and 8 PM. I was not aware that I had to put the time for each day, but my work schedule typically was me stopping when I submitted the work session and working for 75 minutes or more directly before that. To close off the project, I cleaned everything up, tilted all of the problems correctly, and gave the expected solutions when solving the problem.